Revert "Merge tag 'A-Packages'"
This reverts commit 5685e5b81cbc363ba8d153c8249cfa6284c0c9d0, reversing
changes made to 4d8382e39149b5fa65e02b96f338015fd6032ba7.
