+++ src/main/java/Deadline.java
@@ -1,9 +1,7 @@
-package lucian.task;
-
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task {
+public class Deadline extends Task{
 
     protected LocalDate by;
