+++ src/main/java/Event.java
@@ -1,5 +1,3 @@
-package lucian.task;
-
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
