+++ src/main/java/LucianException.java
@@ -1,5 +1,3 @@
-package lucian.exceptions;
-
 public class LucianException extends Exception{
     public LucianException(String string) {
         super(string);
