+++ src/main/java/Task.java
@@ -1,5 +1,3 @@
-package lucian.task;
-
 import java.time.LocalDate;
 
 public abstract class Task {
