+++ src/main/java/ToDo.java
@@ -1,6 +1,4 @@
-package lucian.task;
-
-public class ToDo extends Task {
+public class ToDo extends Task{
 
     public ToDo(String description) {
         super(description);
