+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hey! I'm lucian.Lucian
+Hey! I'm Lucian
 How can I help?
 ____________________________________________________________
 ____________________________________________________________
